refactor(client): rename EditContactModal to Edit UserModal
